app/requirements.txt
SQLAlchemy==1.3.11
sqlalchemy-migrate==0.13.0
Telethon==1.23.0
mysql-connector-python==8.0.18
gspread==3.1.0
oauth2client==4.1.3
python-dotenv==0.19.0